move functions to services
