+++ biofefi/services/ml_models.py
@@ -1,8 +1,19 @@
+import inspect
 import json
 import os
 from pathlib import Path
 from pickle import UnpicklingError, dump, load
 
+from sklearn.ensemble import RandomForestClassifier, RandomForestRegressor
+from sklearn.linear_model import LinearRegression, LogisticRegression
+from sklearn.svm import SVC, SVR
+from xgboost import XGBClassifier, XGBRegressor
+
+from biofefi.machine_learning.nn_models import (
+    BayesianRegularisedNNClassifier,
+    BayesianRegularisedNNRegressor,
+)
+from biofefi.options.enums import ModelNames, ProblemTypes
 from biofefi.utils.utils import create_directory
 
 
@@ -79,4 +90,107 @@ def load_models_to_explain(path: Path, model_names: list) -> dict[str, list]:
                         models[model_name] = [model]
             except UnpicklingError:
                 pass  # ignore bad files
+    return models
+
+
+# Mapping model types and problem types to specific model classes
+_MODEL_PROBLEM_DICT = {
+    (ModelNames.LinearModel, ProblemTypes.Classification): LogisticRegression,
+    (ModelNames.LinearModel, ProblemTypes.Regression): LinearRegression,
+    (ModelNames.RandomForest, ProblemTypes.Classification): RandomForestClassifier,
+    (ModelNames.RandomForest, ProblemTypes.Regression): RandomForestRegressor,
+    (ModelNames.XGBoost, ProblemTypes.Classification): XGBClassifier,
+    (ModelNames.XGBoost, ProblemTypes.Regression): XGBRegressor,
+    (ModelNames.SVM, ProblemTypes.Classification): SVC,
+    (ModelNames.SVM, ProblemTypes.Regression): SVR,
+    (
+        ModelNames.BRNNClassifier,
+        ProblemTypes.Classification,
+    ): BayesianRegularisedNNClassifier,
+    (ModelNames.BRNNRegressor, ProblemTypes.Regression): BayesianRegularisedNNRegressor,
+}
+
+
+def assert_model_param(model, model_params, logger: object = None) -> None:
+    """
+    Asserts that the model parameters are valid
+
+    Parameters
+    ----------
+    model: object
+        The model to check the parameters for
+    model_params: dict
+        The model parameters to check
+    logger: object
+        The logger to use for logging
+    """
+    if hasattr(model, "args"):
+        original_args = model.args
+    else:
+        original_args = list(inspect.signature(model).parameters)
+    args_to_remove = []
+    for arg in model_params:
+        if arg not in original_args:
+            logger.warning(
+                f"Model {model.__name__} does not have parameter {arg}, removing"
+            )
+            args_to_remove.append(arg)
+    for arg in args_to_remove:
+        model_params.pop(arg)
+    # check arguments for XGBoost
+    if "kwargs" in model_params:
+        for arguments, value in model_params["kwargs"].items():
+            model_params[arguments] = value
+        model_params.pop("kwargs")
+
+    logger.info(f"Using model {model.__name__} with parameters {model_params}")
+    return model_params
+
+
+def get_models(
+    model_types: dict[str, dict], problem_type: str, logger: object = None
+) -> list:
+    """
+    Constructs and initializes machine learning models
+    based on the given configuration.
+
+    Args:
+        model_types (dict): Dictionary containing model types
+        and their parameters.
+        problem_type (str): Type of problem (
+            classification or regression).
+        logger (object): Logger object to log messages.
+
+    Returns:
+        - list: A dictionary of initialized models where th
+        keys are model names and the values are instances
+        of the corresponding models.
+
+    Raises:
+        - ValueError: If a model type is not recognized or unsupported
+    """
+    models = {}
+    model_list = [
+        (model_type, model["params"])
+        for model_type, model in model_types.items()
+        if model["use"]
+    ]
+    for model, model_param in model_list:
+        if model_class := _MODEL_PROBLEM_DICT.get(
+            (model.lower(), problem_type.lower())
+        ):
+            if problem_type.lower() == ProblemTypes.Classification:
+                model_param = assert_model_param(
+                    model_class, model_param, logger=logger
+                )
+                model_param["class_weight"] = "balanced"
+                models[model] = model_class(**model_param)
+            else:
+                model_param = assert_model_param(
+                    model_class, model_param, logger=logger
+                )
+                models[model] = model_class(**model_param)
+
+        else:
+            raise ValueError(f"Model type {model} not recognized")
     return models
